# Optimize children cache updates and refine special-case handling