Previously, the Node children cache was marked dirty in many cases, triggering expensive rebuilds of the entire children array even for simple operations like adding or removing a child at the end. This PR refines the cache invalidation logic to handle several special cases.

Changes:
- Adding children at the end no longer marks the cache dirty unnecessarily if the cache is already valid.
- Removing the last child updates the cache without a full rebuild if the cache is already valid.
- Marked Node tree printing and tree string retrieval methods (`print_tree`, `print_tree_pretty`, `get_tree_string`, `get_tree_string_pretty`) as `const` since they do not modify the Node state.
- Changed `children_cache_dirty` default value to `false` so those cases can be handled correctly ie. using `add_child` followed by `get_children` or `get_child_count` will not require rebuilding the cache for those previous special cases.
- Moved `_update_children_cache()` call from `move_child` to `_move_child` to avoid unnecessary calls when the new index is equal the current the index, and also to avoid calling it manually before using `_move_child`. (Note: `move_child` is called on every mouse motion event when rearranging tabs).
- Added a way to recover the HashMap from the cache if it's valid and we fail to remove a child if it's name is not found in the HashMap or we return early before marking the cache dirty which will rebuild the same cache again since the child was not removed from the Hashmap.