@WhalesState 2025-09-18
Script used for benchmark:

```gscript
class_name Control


func _ready() -> void:
	var node: Control = Control.new()
	var internal_count:= 20000
	var count: int = 20000

	for i in internal_count:
		node.add_child(Control.new(), false, Node.INTERNAL_MODE_FRONT)

	for i in count:
		node.add_child(Control.new())

	for i in internal_count:
		node.add_child(Control.new(), false, Node.INTERNAL_MODE_BACK)

	var time = Time.get_ticks_msec()
	node.get_children(true)
	prints(Time.get_ticks_msec() - time, "ms")

	time = Time.get_ticks_msec()
	node.get_children(true)
	prints(Time.get_ticks_msec() - time, "ms")

	node.queue_free()

```
Before: (First calls updates the cache)

```yaml
Before: 47 ms
Before: 20 ms
```

After: Since children are added in order (Back, External, Front) it doesn't require updating the cache.

```yaml
After: 18 ms
After: 16 ms
```