@WhalesState 2025-09-18
There is one potential caveat which is that adding children while updating the cache alongside will be slightly more expensive (because of growth allocations), but I think this is acceptable since we expect children to be queried far more often than added.

We can catch this when it happens and prevent it by adding a debug `mutable bool updating_cache = false;` and an error when it tries to update the cache while it's already updating to know the cases that may invalidate the cache, currently i have found that there are some cases that aren't thread guarded and may cause issues `Node::is_greater_than`, `Node::_get_tree_string_pretty`, `Node::_get_tree_string` and `SceneTreeFTI::_update_dirty_nodes`.

> In any case, the same trade-off was made when originally writing the code, just with fewer cases.

Yes i also have realized that there was a single special case already for adding ordinary children but it wasn't applied for newly added nodes since the cache was marked dirty by default.