@WhalesState 2025-09-21
> I'm not super expert on the children cache in 4.x (I work mainly with 3.x which doesn't have this), but the PR looks fine from a quick look, assuming initializing `dirty` to false is okay (it depends how it is initially populated, sometimes there are shortcuts for loading, so there is just a possibility of regressions there worth checking).

The cache is only marked dirty when children are added, removed, or moved. So initializing it to false is safe by default, those operations will properly detect and update the dirty state when needed.